fixes & code refactored
