+++ mochaccino-client/src/Components/Modal/Modal.css
@@ -10,10 +10,4 @@
   justify-content: center;
   flex-wrap: wrap;
   align-items: center;
-}
-
-.modal__title-choosed-wrapper {
-  display: flex;
-  align-items: center;
-  justify-content: flex-start;
 }/*# sourceMappingURL=Modal.css.map */
