+++ mochaccino-client/src/Components/Modal/Modal.scss
@@ -9,8 +9,3 @@
   flex-wrap: wrap;
   align-items: center;
 }
-.modal__title-choosed-wrapper {
-  display: flex;
-  align-items: center;
-  justify-content: flex-start;
-}
