+++ mochaccino-client/src/Components/Spinner/SpinnerComponent.css
@@ -1,0 +1,7 @@
+.spinner-component-wrapper {
+  width: 100%;
+  height: 150px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}/*# sourceMappingURL=SpinnerComponent.css.map */
